+++ src/main/java/com/photocontest/dao/impl/VoterDAOImpl.java
@@ -14,7 +14,7 @@ import javax.persistence.Query;
  * Time: 2:22 AM
  * To change this template use File | Settings | File Templates.
  */
-public class VoterDAOImpl extends GenericDAOImpl<Voter, Integer> implements VoterDAO {
+public class VoterDAOImpl extends GenericDAOImpl<Voter, Long> implements VoterDAO {
 
     public VoterDAOImpl(){
         super(Voter.class);
